separate bundleManager.js
